+++ src/irvine/oeis/a001/A001692.java
@@ -43,7 +43,7 @@ public class A001692 extends AbstractSequence {
     Z s = Z.ZERO;
     for (final Z d : Jaguar.factor(mN).divisors()) {
       final int dd = d.intValue();
-      s = s.add(mPowers.get(mN / dd).multiply(Functions.MOBIUS.i((long) dd)));
+      s = s.add(mPowers.get(mN / dd).multiply(Functions.MOBIUS.i(dd)));
     }
     return s.divide(mN);
   }
