+++ src/irvine/oeis/a006/A006575.java
@@ -20,7 +20,7 @@ public class A006575 extends Sequence1 {
     for (final Z dd : Jaguar.factor(mN).divisors()) {
       final int d = dd.intValue();
       if ((d & 1) == 1) {
-        sum = sum.add(Z.THREE.pow(mN / d).subtract(1).multiply(Functions.MOBIUS.i((long) d)));
+        sum = sum.add(Z.THREE.pow(mN / d).subtract(1).multiply(Functions.MOBIUS.i(d)));
       }
     }
     return sum.divide(2L * mN);
