+++ src/irvine/oeis/a034/A034747.java
@@ -18,7 +18,7 @@ public class A034747 extends Sequence1 {
     Z sum = Z.ZERO;
     for (final Z dd : Jaguar.factor(++mN).divisors()) {
       final int d = dd.intValue();
-      sum = sum.add(Functions.SIGMA1.z(mN / d).multiply(Functions.FIBONACCI.z((long) d)));
+      sum = sum.add(Functions.SIGMA1.z(mN / d).multiply(Functions.FIBONACCI.z(d)));
     }
     return sum;
   }
