+++ src/irvine/oeis/a034/A034748.java
@@ -18,7 +18,7 @@ public class A034748 extends Sequence1 {
     Z sum = Z.ZERO;
     for (final Z dd : Jaguar.factor(++mN).divisors()) {
       final int d = dd.intValue();
-      sum = sum.add(Functions.PHI.z((long) mN / d).multiply(Functions.FIBONACCI.z((long) d)));
+      sum = sum.add(Functions.PHI.z(mN / d).multiply(Functions.FIBONACCI.z(d)));
     }
     return sum;
   }
